+++ apps/server/src/utils/business.ts
@@ -1,11 +1,15 @@
+import { PluginResult } from '@/plugins/hooks/hook.js'
+import { hooks } from '@/plugins/index.js'
 import {
   getEnvironmentsByProjectId,
   getProjectRepositories,
   getProjectById,
   lockProject,
   unlockProject,
+  addLogs,
 } from '@/resources/queries-index.js'
-import type { Project } from '@prisma/client'
+import type { Project, User } from '@prisma/client'
+import { BadRequestError } from './errors.js'
 
 export const unlockProjectIfNotFailed = async (projectId: Project['id']) => {
   const environmentStatuses = (await getEnvironmentsByProjectId(projectId))?.map(environment => environment?.status)
@@ -20,5 +24,21 @@ export const unlockProjectIfNotFailed = async (projectId: Project['id']) => {
     return lockProject(projectId)
   } else {
     return unlockProject(projectId)
+  }
+}
+
+export const checkCreateProject = async (owner: User, resource: 'Project' | 'Repository') => {
+  const pluginsResults = await hooks.createProject.validate({ owner })
+  if (pluginsResults?.failed) {
+    const reasons = Object.values(pluginsResults)
+      .filter((plugin: PluginResult) => plugin?.status?.result === 'KO')
+      .map((plugin: PluginResult) => plugin.status.message)
+      .join('; ')
+
+    // @ts-ignore
+    await addLogs(`Create ${resource} Validation`, pluginsResults, owner.id)
+
+    const message = 'Echec de la validation des prérequis par les services externes'
+    throw new BadRequestError(message, { description: reasons })
   }
 }
